MOBILE-3947: Upgrade to Ionic 7

+++ src/testing/services/behat-dom.ts
@@ -48,7 +48,7 @@ export class TestingBehatDomUtilsService {
             return false;
         }
 
-        if (element.tagName === 'ION-SLIDE') {
+        if (element.tagName === 'SWIPER-SLIDE') {
             // Check if the slide is visible (in the viewport).
             const bounding = element.getBoundingClientRect();
             if (bounding.right <= 0 || bounding.left >= window.innerWidth) {
@@ -76,25 +76,38 @@ export class TestingBehatDomUtilsService {
      * Check if an element is selected.
      *
      * @param element Element.
-     * @param container Container.
+     * @param firstCall Whether this is the first call of the function.
      * @returns Whether the element is selected or not.
      */
-    isElementSelected(element: HTMLElement, container: HTMLElement): boolean {
+    isElementSelected(element: HTMLElement, firstCall = true): boolean {
         const ariaCurrent = element.getAttribute('aria-current');
-        if (
-            (ariaCurrent && ariaCurrent !== 'false') ||
-            (element.getAttribute('aria-selected') === 'true') ||
-            (element.getAttribute('aria-checked') === 'true')
-        ) {
-            return true;
+        const ariaSelected = element.getAttribute('aria-selected');
+        const ariaChecked = element.getAttribute('aria-checked');
+
+        if (ariaCurrent || ariaSelected || ariaChecked) {
+            return (!!ariaCurrent && ariaCurrent !== 'false') ||
+                (!!ariaSelected && ariaSelected === 'true') ||
+                (!!ariaChecked && ariaChecked === 'true');
+        }
+
+        if (firstCall) {
+            const inputElement =  element.closest('ion-checkbox, ion-radio, ion-toggle')?.querySelector('input');
+            if (inputElement) {
+                return inputElement.value === 'on';
+            }
+
+            const tabButtonElement =  element.closest('ion-tab-button');
+            if (tabButtonElement?.classList.contains('tab-selected')) {
+                return true;
+            }
         }
 
         const parentElement = this.getParentElement(element);
-        if (!parentElement || parentElement === container) {
+        if (!parentElement || parentElement.classList.contains('ion-page')) {
             return false;
         }
 
-        return this.isElementSelected(parentElement, container);
+        return this.isElementSelected(parentElement, false);
     }
 
     /**
@@ -263,7 +276,7 @@ export class TestingBehatDomUtilsService {
     /**
      * Given a list of elements, get the top ancestors among all of them.
      *
-     * This will remote duplicates and drop any elements nested within each other.
+     * This will remove duplicates and drop any elements nested within each other.
      *
      * @param elements Elements list.
      * @returns Top ancestors.
@@ -277,7 +290,18 @@ export class TestingBehatDomUtilsService {
                     continue;
                 }
 
-                if (element.contains(otherElement)) {
+                let documentPosition = element.compareDocumentPosition(otherElement);
+                // eslint-disable-next-line no-bitwise
+                if (documentPosition & Node.DOCUMENT_POSITION_DISCONNECTED) {
+                    // Check if they are inside shadow DOM so we can compare their hosts.
+                    const elementHost = this.getShadowDOMHost(element) || element;
+                    const otherElementHost = this.getShadowDOMHost(otherElement) || otherElement;
+
+                    documentPosition = elementHost.compareDocumentPosition(otherElementHost);
+                }
+
+                // eslint-disable-next-line no-bitwise
+                if (documentPosition & Node.DOCUMENT_POSITION_CONTAINS) {
                     uniqueElements.delete(otherElement);
                 }
             }
@@ -293,9 +317,22 @@ export class TestingBehatDomUtilsService {
      * @returns Parent element.
      */
     protected getParentElement(element: HTMLElement): HTMLElement | null {
-        return element.parentElement ||
-            (element.getRootNode() && (element.getRootNode() as ShadowRoot).host as HTMLElement) ||
-            null;
+        return element.parentElement || this.getShadowDOMHost(element);
+    }
+
+    /**
+     * Get shadow DOM host element.
+     *
+     * @param element Element.
+     * @returns Shadow DOM host element.
+     */
+    protected getShadowDOMHost(element: HTMLElement): HTMLElement | null {
+        const node = element.getRootNode();
+        if (node instanceof ShadowRoot) {
+            return node.host as HTMLElement;
+        }
+
+        return null;
     }
 
     /**
@@ -391,10 +428,17 @@ export class TestingBehatDomUtilsService {
      * @returns Field element.
      */
     findField(field: string): HTMLElement | HTMLInputElement | undefined {
-        const input = this.findElementBasedOnText(
-            { text: field, selector: 'input, textarea, [contenteditable="true"], ion-select, ion-datetime' },
+        const selector =
+            'input, textarea, core-rich-text-editor, [contenteditable="true"], ion-select, ion-datetime-button, ion-datetime';
+
+        let input = this.findElementBasedOnText(
+            { text: field, selector },
             { onlyClickable: false, containerName: '' },
         );
+
+        if (input?.tagName === 'CORE-RICH-TEXT-EDITOR') {
+            input = input.querySelector<HTMLElement>('[contenteditable="true"]') || undefined;
+        }
 
         if (input) {
             return input;
@@ -408,7 +452,17 @@ export class TestingBehatDomUtilsService {
         if (label) {
             const inputId = label.getAttribute('for');
 
-            return (inputId && document.getElementById(inputId)) || undefined;
+            if (inputId) {
+                return document.getElementById(inputId) || undefined;
+            }
+
+            input = this.getShadowDOMHost(label) || undefined;
+
+            // Add support for other input types if required by adding them to the array.
+            const ionicInputFields = ['ION-INPUT', 'ION-TEXTAREA', 'ION-SELECT', 'ION-DATETIME', 'ION-TOGGLE'];
+            if (input && ionicInputFields.includes(input.tagName)) {
+                return input;
+            }
         }
     }
 
@@ -424,6 +478,34 @@ export class TestingBehatDomUtilsService {
         options: TestingBehatFindOptions = {},
     ): HTMLElement | undefined {
         return this.findElementsBasedOnText(locator, options)[0];
+    }
+
+    /**
+     * Wait until an element with the given selector is found.
+     *
+     * @param selector Element selector.
+     * @param timeout Timeout after which an error is thrown.
+     * @param retryFrequency Frequency for retries when the element is not found.
+     * @returns Element.
+     */
+    async waitForElement<T extends HTMLElement = HTMLElement>(
+        selector: string,
+        timeout: number = 2000,
+        retryFrequency: number = 100,
+    ): Promise<T> {
+        const element = document.querySelector<T>(selector);
+
+        if (!element) {
+            if (timeout < retryFrequency) {
+                throw new Error(`Element with '${selector}' selector not found`);
+            }
+
+            await new Promise(resolve => setTimeout(resolve, retryFrequency));
+
+            return this.waitForElement<T>(selector, timeout - retryFrequency, retryFrequency);
+        }
+
+        return element;
     }
 
     /**
@@ -461,7 +543,7 @@ export class TestingBehatDomUtilsService {
     protected findElementsBasedOnTextInContainer(
         locator: TestingBehatElementLocator,
         topContainer: HTMLElement,
-        options: TestingBehatFindOptions,
+        options: TestingBehatFindOptions = {},
     ): HTMLElement[] {
         let container: HTMLElement | null = topContainer;
 
@@ -579,8 +661,10 @@ export class TestingBehatDomUtilsService {
             // may not work without doing this.
             const parentElement = this.getParentElement(element);
 
-            if (parentElement && parentElement.matches('ion-button, ion-back-button')) {
+            if (parentElement?.matches('ion-button, ion-back-button')) {
                 element = parentElement;
+            } else if (parentElement?.tagName === 'ION-ITEM' && parentElement?.classList.contains('clickable')) {
+                element = parentElement.querySelector<HTMLElement>('ion-toggle') || element;
             }
 
             const rect = await this.ensureElementVisible(element);
@@ -611,31 +695,26 @@ export class TestingBehatDomUtilsService {
     }
 
     /**
-     * Set an element value.
+     * Set an input element value.
      *
-     * @param element HTML to set.
-     * @param value Value to be set.
+     * @param element Input element.
+     * @param value Value.
      */
-    async setElementValue(element: HTMLInputElement | HTMLElement, value: string): Promise<void> {
+    async setInputValue(element: HTMLInputElement | HTMLElement, value: string): Promise<void> {
         await NgZone.run(async () => {
-            const promise = new CorePromisedValue<void>();
-
             // Functions to get/set value depending on field type.
-            const setValue = (text: string) => {
-                if (element.tagName === 'ION-SELECT' && 'value' in element) {
-                    value = value.trim();
-                    const optionValue = Array.from(element.querySelectorAll('ion-select-option'))
-                        .find((option) => option.innerHTML.trim() === value);
-
-                    if (optionValue) {
-                        element.value = optionValue.value;
-                    }
+            const setValue = async (text: string) => {
+                if (element.tagName === 'ION-SELECT') {
+                    this.setIonSelectInputValue(element, value);
                 } else if ('value' in element) {
                     element.value = text;
                 } else {
                     element.innerHTML = text;
                 }
+
+                element.dispatchEvent(new Event('ionChange'));
             };
+
             const getValue = () => {
                 if ('value' in element) {
                     return element.value;
@@ -645,38 +724,79 @@ export class TestingBehatDomUtilsService {
             };
 
             // Pretend we have cut and pasted the new text.
-            let event: InputEvent;
-            if (getValue() !== '') {
-                event = new InputEvent('input', {
+            if (element.tagName !== 'ION-SELECT' && getValue() !== '') {
+                await CoreUtils.nextTick();
+                await setValue('');
+
+                element.dispatchEvent(new InputEvent('input', {
                     bubbles: true,
                     view: window,
                     cancelable: true,
                     inputType: 'deleteByCut',
-                });
-
-                await CoreUtils.nextTick();
-                setValue('');
-                element.dispatchEvent(event);
+                }));
             }
 
             if (value !== '') {
-                event = new InputEvent('input', {
+                await CoreUtils.nextTick();
+                await setValue(value);
+
+                element.dispatchEvent(new InputEvent('input', {
                     bubbles: true,
                     view: window,
                     cancelable: true,
                     inputType: 'insertFromPaste',
                     data: value,
-                });
+                }));
+            }
+        });
+    }
 
-                await CoreUtils.nextTick();
-                setValue(value);
-                element.dispatchEvent(event);
+    /**
+     * Select an option in an ion-select element.
+     *
+     * @param element IonSelect element.
+     * @param value Value.
+     */
+    protected async setIonSelectInputValue(element: HTMLElement, value: string): Promise<void> {
+        // Press select.
+        await TestingBehatDomUtils.pressElement(element);
+
+        // Press option.
+        type IonSelectInterface = 'alert' | 'action-sheet' | 'popover';
+        const selectInterface = element.getAttribute('interface') as IonSelectInterface ?? 'alert';
+        const containerSelector = ({
+            'alert': 'ion-alert.select-alert',
+            'action-sheet': 'ion-action-sheet.select-action-sheet',
+            'popover': 'ion-popover.select-popover',
+        })[selectInterface];
+        const optionSelector = ({
+            'alert': 'button',
+            'action-sheet': 'button',
+            'popover': 'ion-radio',
+        })[selectInterface] ?? '';
+        const optionsContainer = await TestingBehatDomUtils.waitForElement(containerSelector);
+        const options = this.findElementsBasedOnTextInContainer(
+            { text: value, selector: optionSelector },
+            optionsContainer,
+            {},
+        );
+
+        if (options.length === 0) {
+            throw new Error('Couldn\'t find ion-select option.');
+        }
+
+        await TestingBehatDomUtils.pressElement(options[0]);
+
+        // Press options submit.
+        if (selectInterface === 'alert') {
+            const submitButton = optionsContainer.querySelector<HTMLElement>('.alert-button-group button:last-child');
+
+            if (!submitButton) {
+                throw new Error('Couldn\'t find ion-select submit button.');
             }
 
-            promise.resolve();
-
-            return promise;
-        });
+            await TestingBehatDomUtils.pressElement(submitButton);
+        }
     }
 
 }
